requirements.txt
web3>=6.0.0
aiohttp>=3.9.0
aiosqlite>=0.19.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
eth-account>=0.10.0
pytest>=7.4.0
pytest-asyncio>=0.23.0
